requirements.txt
gradio
pillow
SpeechRecognition
